+++ sodata-client/src/main/java/ch/so/agi/sodata/MapPresets.java
@@ -1,0 +1,111 @@
+package ch.so.agi.sodata;
+
+import ol.Collection;
+import ol.Coordinate;
+import ol.Extent;
+import ol.Map;
+import ol.MapOptions;
+import ol.OLFactory;
+import ol.View;
+import ol.ViewOptions;
+import ol.control.Control;
+import ol.interaction.DefaultInteractionsOptions;
+import ol.interaction.Interaction;
+import ol.layer.Base;
+import ol.layer.Group;
+import ol.layer.Image;
+import ol.layer.LayerOptions;
+import ol.layer.Tile;
+import ol.proj.Projection;
+import ol.proj.ProjectionOptions;
+import ol.source.ImageWms;
+import ol.source.ImageWmsOptions;
+import ol.source.ImageWmsParams;
+import ol.source.TileWms;
+import ol.source.TileWmsOptions;
+import ol.source.TileWmsParams;
+import ol.source.Wmts;
+import ol.source.WmtsOptions;
+import ol.tilegrid.TileGrid;
+import ol.tilegrid.WmtsTileGrid;
+import ol.tilegrid.WmtsTileGridOptions;
+import proj4.Proj4;
+
+public class MapPresets {
+    
+    private MapPresets() {
+        throw new AssertionError();
+    }
+
+    public static double resolutions[] = new double[] { 4000.0, 2000.0, 1000.0, 500.0, 250.0, 100.0, 50.0, 20.0, 10.0, 5.0, 2.5, 1.0, 0.5, 0.25, 0.1 };
+
+    public static Map getColorMap(String mapId) {
+        Proj4.defs("EPSG:2056", "+proj=somerc +lat_0=46.95240555555556 +lon_0=7.439583333333333 +k_0=1 +x_0=2600000 +y_0=1200000 +ellps=bessel +towgs84=674.374,15.056,405.346,0,0,0,0 +units=m +no_defs");
+
+        ProjectionOptions projectionOptions = OLFactory.createOptions();
+        projectionOptions.setCode("EPSG:2056");
+        projectionOptions.setUnits("m");
+        projectionOptions.setExtent(new Extent(2420000, 1030000, 2900000, 1350000));
+        Projection projection = new Projection(projectionOptions);
+        Projection.addProjection(projection);
+
+        WmtsOptions wmtsOptions = OLFactory.createOptions();
+        wmtsOptions.setUrl("https://geo.so.ch/api/wmts/1.0.0/{Layer}/default/2056/{TileMatrix}/{TileRow}/{TileCol}");
+        wmtsOptions.setLayer("ch.so.agi.hintergrundkarte_sw");
+        wmtsOptions.setRequestEncoding("REST");
+        wmtsOptions.setFormat("image/png");
+        wmtsOptions.setMatrixSet("EPSG:2056");
+        wmtsOptions.setStyle("default");
+        wmtsOptions.setProjection(projection);
+        wmtsOptions.setWrapX(true);
+        wmtsOptions.setTileGrid(createWmtsTileGrid(projection, resolutions));
+
+        Wmts wmtsSource = new Wmts(wmtsOptions);
+
+        LayerOptions wmtsLayerOptions = OLFactory.createOptions();
+        wmtsLayerOptions.setSource(wmtsSource);
+
+        Tile wmtsLayer = new Tile(wmtsLayerOptions);
+        wmtsLayer.setOpacity(1);
+
+        ViewOptions viewOptions = OLFactory.createOptions();
+        viewOptions.setProjection(projection);
+        viewOptions.setResolutions(new double[] { 4000.0, 2000.0, 1000.0, 500.0, 250.0, 100.0, 50.0, 20.0, 10.0, 5.0, 2.5, 1.0, 0.5, 0.25, 0.1 });
+        View view = new View(viewOptions);
+        Coordinate centerCoordinate = new Coordinate(2616491, 1240287);
+
+        view.setCenter(centerCoordinate);
+        view.setZoom(6);
+        
+        MapOptions mapOptions = OLFactory.createOptions();
+        mapOptions.setTarget(mapId);
+        mapOptions.setView(view);
+        mapOptions.setControls(new Collection<Control>());
+
+        DefaultInteractionsOptions interactionOptions = new ol.interaction.DefaultInteractionsOptions();
+        interactionOptions.setPinchRotate(false);
+        mapOptions.setInteractions(Interaction.defaults(interactionOptions));
+
+        Map map = new Map(mapOptions);
+        map.addLayer(wmtsLayer);
+        
+        return map;
+    }
+    
+    private static TileGrid createWmtsTileGrid(Projection projection, double[] resolutions) {
+        WmtsTileGridOptions wmtsTileGridOptions = OLFactory.createOptions();
+        
+        String[] matrixIds = new String[resolutions.length];
+
+        for (int z = 0; z < resolutions.length; ++z) {
+            matrixIds[z] = String.valueOf(z);
+        }
+
+        Coordinate tileGridOrigin = projection.getExtent().getTopLeft();
+        wmtsTileGridOptions.setOrigin(tileGridOrigin);
+        wmtsTileGridOptions.setResolutions(resolutions);
+        wmtsTileGridOptions.setMatrixIds(matrixIds);
+
+        return new WmtsTileGrid(wmtsTileGridOptions);
+    }
+}
