+++ sodata-client/src/main/java/ch/so/agi/sodata/MyMessages.java
@@ -1,0 +1,11 @@
+package ch.so.agi.sodata;
+
+import com.google.gwt.i18n.client.Messages;
+
+public interface MyMessages extends Messages {
+    @DefaultMessage("Fubar")
+    String fubar();
+    
+    @DefaultMessage("Füü {0} bar {1}")
+    String yinyang(String yin, String yang);
+}
