+++ sodata-client/src/main/java/ch/so/agi/sodata/MyMessages_de.properties
