+++ sodata-client/src/main/module.gwt.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module>
+  <inherits name="com.google.gwt.user.User" />
+  <inherits name="com.google.gwt.json.JSON" />
+  <inherits name="com.google.gwt.i18n.I18N" />
+
+  <extend-property name="locale" values="de_CH" />
+  <set-property name="locale" value="de_CH" />
+    
+  <inherits name="org.dominokit.domino.ui.DominoUI" />
+  <inherits name="ol.GwtOL" />
+  <inherits name="proj4.GwtProj4" />
+        
+  <source path="" />
+
+  <entry-point class="ch.so.agi.sodata.App" />
+
+  <!-- Only support recent browsers -->
+  <set-property name="user.agent" value="ie10,gecko1_8,safari" />
+  
+  <inherits name="com.google.gwt.core.Core" />
+  <add-linker name="direct_install" />  
+</module>
