+++ sodata-server/src/main/docker/Dockerfile.jvm
@@ -1,0 +1,19 @@
+FROM bellsoft/liberica-openjre-alpine:11.0.12-7 as builder
+WORKDIR /home/app
+ARG JAR_FILE=sodata-server/target/sodata.jar
+COPY ${JAR_FILE} /home/app/application.jar
+RUN java -Djarmode=layertools -jar /home/app/application.jar extract
+
+FROM bellsoft/liberica-openjre-alpine:11.0.12-7
+EXPOSE 8080
+WORKDIR /home/app
+COPY --from=builder /home/app/dependencies/ ./
+COPY --from=builder /home/app/spring-boot-loader/ ./
+COPY --from=builder /home/app/snapshot-dependencies/ ./
+RUN true 
+COPY --from=builder /home/app/application/ ./
+RUN chown -R 1001:0 /home/app && \
+    chmod -R g=u /home/app
+USER 1001
+ENTRYPOINT ["java" ,"-XX:MaxRAMPercentage=80.0", "-noverify", "-XX:TieredStopAtLevel=1", "org.springframework.boot.loader.JarLauncher"]
+
