+++ sodata-server/src/main/docker/Dockerfile.native
@@ -1,0 +1,12 @@
+FROM registry.access.redhat.com/ubi8/ubi-minimal:8.3
+
+WORKDIR /work/
+RUN chown 1001 /work \
+    && chmod "g+rwX" /work \
+    && chown 1001:root /work
+COPY --chown=1001:root sodata-server/target/sodata-server /work/application
+
+EXPOSE 8080
+USER 1001
+
+CMD ["./application"]
