+++ sodata-server/src/main/docker/Dockerfile.native-build
@@ -1,0 +1,26 @@
+FROM oracle/graalvm-ce:20.3.0-java11 
+RUN gu install native-image
+RUN native-image --version
+COPY . /home/app
+WORKDIR /home/app
+RUN ./mvnw -e -q -B -Penv-prod,native clean package && \
+    ls -la /home/app/sodata-server
+
+FROM debian:buster-slim
+ADD https://github.com/upx/upx/releases/download/v3.96/upx-3.96-amd64_linux.tar.xz /usr/local
+RUN apt-get update && apt-get install -y xz-utils && \
+    xz -d -c /usr/local/upx-3.96-amd64_linux.tar.xz | tar -xOf - upx-3.96-amd64_linux/upx > /bin/upx && \
+    chmod a+x /bin/upx
+WORKDIR /home/app
+COPY --from=0 /home/app/sodata-server/target/ch.so.agi.sodata.application /home/app
+RUN upx -7 -k /home/app/ch.so.agi.sodata.application
+
+FROM registry.access.redhat.com/ubi8/ubi-minimal:8.3
+WORKDIR /work/
+RUN chown 1001 /work \
+    && chmod "g+rwX" /work \
+    && chown 1001:root /work
+COPY --from=1 --chown=1001:root /home/app/ch.so.agi.sodata.application /work/application
+EXPOSE 8080
+USER 1001
+CMD ["./application"]
