+++ sodata-server/src/main/java/ch/so/agi/sodata/Application.java
@@ -1,0 +1,22 @@
+package ch.so.agi.sodata;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.ServletComponentScan;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+@ServletComponentScan
+public class Application extends SpringBootServletInitializer {
+  
+  public static void main(String[] args) {
+    SpringApplication.run(Application.class,
+                          args);
+  }
+  
+  @Override
+  protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
+    return builder.sources(Application.class);
+  }
+}
