+++ sodata-server/src/main/java/ch/so/agi/sodata/MainController.java
@@ -1,0 +1,22 @@
+package ch.so.agi.sodata;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MainController {
+    private final Logger log = LoggerFactory.getLogger(this.getClass());
+
+    @Autowired
+    Settings settings;
+    
+    @GetMapping("/settings")
+    public ResponseEntity<?> getSettings() {
+        return ResponseEntity.ok().body(settings);
+    }
+}
