+++ sodata-server/src/main/java/ch/so/agi/sodata/Settings.java
@@ -1,0 +1,20 @@
+package ch.so.agi.sodata;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@EnableConfigurationProperties
+@ConfigurationProperties(prefix = "app")
+public class Settings {
+    private String myVar;
+
+    public String getMyVar() {
+        return myVar;
+    }
+
+    public void setMyVar(String myVar) {
+        this.myVar = myVar;
+    }
+}
