+++ sodata-server/src/main/resources/META-INF/native-image/resource-config.json
@@ -1,0 +1,18 @@
+{
+  "resources":{
+  "includes":[
+    {"pattern":"\\Qpublic/index.html\\E"},
+    {"pattern":"\\Qpublic/icon.png\\E"},
+    {"pattern":"\\Qpublic/app/font/MaterialIcons-Regular.eot\\E"},
+    {"pattern":"\\Qpublic/app/font/MaterialIcons-Regular.ttf\\E"},
+    {"pattern":"\\Qpublic/app/font/MaterialIcons-Regular.woff\\E"},
+    {"pattern":"\\Qpublic/app/font/MaterialIcons-Regular.woff2\\E"},
+    {"pattern":"\\Qpublic/app/css/mdi/fonts/materialdesignicons-webfont.eot\\E"},    
+    {"pattern":"\\Qpublic/app/css/mdi/fonts/materialdesignicons-webfont.ttf\\E"},    
+    {"pattern":"\\Qpublic/app/css/mdi/fonts/materialdesignicons-webfont.woff\\E"},    
+    {"pattern":"\\Qpublic/app/css/mdi/fonts/materialdesignicons-webfont.woff2\\E"},        
+    {"pattern": ".*\\.js$"},
+    {"pattern": ".*\\.css$"}        
+  ]},
+  "bundles":[]
+}
