+++ sodata-server/src/main/resources/application-prod.yml
@@ -1,0 +1,8 @@
+# ========================
+# SPRING
+# ========================
+spring:
+    devtools:
+        add-properties: false
+        livereload:
+            enabled: false
