+++ sodata-server/src/main/resources/application.yml
@@ -1,0 +1,34 @@
+# ========================
+# TOMCAT
+# ========================
+server:
+  tomcat:
+    additional-tld-skip-patterns: '*.jar'
+  compression:
+    enabled: true
+    mime-types: text/html,text/xml,text/plain,text/css,text/javascript,application/javascript,application/json
+    min-response-size: 1024
+
+# ========================
+# SPRING
+# ========================
+logging:
+  level:
+    org.springframework: INFO 
+
+spring:
+    resources:
+        static-locations: classpath:/launcherDir/,classpath:/public/
+    devtools:
+        restart:
+            additional-exclude: launcherDir/**
+
+# spring:
+#     profiles:
+#         active: test 
+
+# ========================
+# APPLICATION
+# ========================
+app:
+    myVar: ${VAR:defaultValue}
