+++ sodata-server/src/main/resources/public/android-chrome-192x192.png
