+++ sodata-server/src/main/resources/public/favicon-16x16.png
