+++ sodata-server/src/main/resources/public/favicon-32x32.png
