+++ sodata-server/src/main/resources/public/favicon-96x96.png
