+++ sodata-server/src/main/resources/public/favicon.ico
