+++ sodata-server/src/main/resources/public/fonts.css
