+++ sodata-shared/src/main/java/ch/so/agi/sodata/Empty.java
@@ -1,0 +1,3 @@
+package ch.so.agi.sodata;
+
+class Empty {}
