--- src/main/java/ch/so/agi/sodata/BootGwtApplication.java
@@ -1,32 +1,0 @@
-package ch.so.agi.sodata;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.ServletComponentScan;
-import org.springframework.boot.web.servlet.ServletRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.filter.ForwardedHeaderFilter;
-
-import ch.so.agi.sodata.server.SettingsServiceImpl;
-
-@ServletComponentScan
-@SpringBootApplication
-@Configuration
-public class BootGwtApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(BootGwtApplication.class, args);
-	}
-	
-    @Bean
-    public ForwardedHeaderFilter forwardedHeaderFilter() {
-        return new ForwardedHeaderFilter();
-    }
-	
-    @Bean
-    public ServletRegistrationBean configServletBean() {
-        ServletRegistrationBean bean = new ServletRegistrationBean(new SettingsServiceImpl(), "/module1/settings");
-        bean.setLoadOnStartup(1);
-        return bean;
-    }
-}
