--- src/main/java/ch/so/agi/sodata/GWTCacheControlFilter.java
@@ -1,38 +1,0 @@
-package ch.so.agi.sodata;
-
-import java.io.IOException;
-import java.util.Date;
-
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class GWTCacheControlFilter implements Filter {
-
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
-            throws IOException, ServletException {
-
-        HttpServletRequest httpRequest = (HttpServletRequest) request;
-        String requestURI = httpRequest.getRequestURI();
-
-        if (requestURI.contains(".nocache.")) {
-            Date now = new Date();
-            HttpServletResponse httpResponse = (HttpServletResponse) response;
-            httpResponse.setDateHeader("Date", now.getTime());
-            // one day old
-            httpResponse.setDateHeader("Expires", now.getTime() - 86400000L);
-            httpResponse.setHeader("Pragma", "no-cache");
-            httpResponse.setHeader("Cache-control", "no-cache, no-store, must-revalidate");
-        }
-        chain.doFilter(request, response);
-    }
-
-}
