--- src/main/java/ch/so/agi/sodata/client/MyMessages.java
@@ -1,11 +1,0 @@
-package ch.so.agi.sodata.client;
-
-import com.google.gwt.i18n.client.Messages;
-
-public interface MyMessages extends Messages {
-    @DefaultMessage("Fubar")
-    String fubar();
-    
-    @DefaultMessage("Füü {0} bar {1}")
-    String yinyang(String yin, String yang);
-}
