--- src/main/java/ch/so/agi/sodata/server/search/LuceneSearcherException.java
@@ -1,14 +1,0 @@
-package ch.so.agi.sodata.server.search;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value=HttpStatus.INTERNAL_SERVER_ERROR)
-public class LuceneSearcherException extends Exception {
-
-    private static final long serialVersionUID = 1L;
-
-    public LuceneSearcherException(String msg) {
-        super(msg);
-    }
-}
