--- src/main/java/ch/so/agi/sodata/shared/SettingsResponse.java
@@ -1,16 +1,0 @@
-package ch.so.agi.sodata.shared;
-
-import java.util.HashMap;
-import com.google.gwt.user.client.rpc.IsSerializable;
-
-public class SettingsResponse implements IsSerializable {
-    private HashMap<String,Object> settings;
-
-    public HashMap<String, Object> getSettings() {
-        return settings;
-    }
-
-    public void setSettings(HashMap<String, Object> settings) {
-        this.settings = settings;
-    } 
-}
