--- src/main/java/ch/so/agi/sodata/shared/SettingsServiceAsync.java
@@ -1,8 +1,0 @@
-package ch.so.agi.sodata.shared;
-
-import com.google.gwt.user.client.rpc.AsyncCallback;
-
-public interface SettingsServiceAsync {
-    void settingsServer(AsyncCallback<SettingsResponse> callback)
-            throws IllegalArgumentException;
-}
