--- src/main/resources/static/client.css
@@ -1,213 +1,0 @@
-body {
-    margin: 0px;
-    padding: 0px;
-    background: #FFFFFF; 
-}
-
-body, div {
-    /*font-family: 'Fira Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;*/
-    font-family: Frutiger, 'Noto Sans JP', 'Barlow', 'Fira Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
-    font-weight: 300;
-}
-
-#map {
-    width: 100%;
-    height: 600px;
-}
-
-#map-plus-row {
-    margin-top: 30px;
-}
-
-.field-group input, .field-group textarea {
-    /*font-family: 'Fira Sans', sans-serif !important;*/   
-    font-family: Frutiger, 'Noto Sans JP', 'Barlow', 'Fira Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif !important;   
-    font-size: 16px;   
-    font-weight: 300;     
-}
-
-.field-group .field-label {
-	font-size: 12px;
-    font-weight: 300;	
-}
-
-.dropdown-menu>li>a {
-	font-weight: 300;
-} 
-
-#container {
-    margin-left: auto;
-    margin-right: auto;
-    width: 1400px;
-    max-width: 95%;
-    background-color: #FFFFFF;        
-}
-
-.logo {
-    margin-top: 10px;
-    font-size: 40px;
-    text-align: right;
-}
-
-.breadcrumb {
-    padding: 0px !important;	
-}
-
-.sodata-title {
-    margin-top: 120px;
-    font-size: 40px;
-}
-
-.info {
-    margin-top: 30px;
-    font-size: 16px;
-    font-style: normal;
-}
-
-#search-panel {
-    margin-top: 40px;
-    margin-bottom: 30px;
-}
-
-/*
-a {
-    color: #ef5350;  
-}
-
-a:hover {
-    color: #ef5350;
-    text-decoration: underline !important;
-    text-decoration-thickness: 2px !important;   
-}  
-
-a:visited {
-    color: #ef5350; 
-    text-decoration: underline !important;
-} 
-*/
-
-.bg-white.nav-tabs li a {
-	color: #333333;
-}
-
-.bg-white.nav-tabs li a, .bg-white.nav-tabs li a:hover {
-    color: #333333;
-}
-
-h4, .h4 {
-    font-size: 16px !important;
-}
-
-a {
-    color: #333333; 
-    /*text-decoration: underline !important;*/
-}
-
-a:hover {
-    /*color: #ef5350;*/
-    /*text-decoration: underline !important;*/
-}  
-
-a.generic-link {
-    color: #333333;
-    text-decoration: underline !important;
-} 
-
-a.generic-link:hover {
-    color: #ef5350;
-    text-decoration: underline !important;
-}  
-
-a:visited {
-    color: #333333; 
-    /*text-decoration: underline !important;*/
-}
-
-.bg-l-4-grey {
-    color: #999999 !important;
-    font-weight: 300 !important;
-}
-
-.bg-l-3-grey {
-    color: #666666 !important;
-    font-weight: 300 !important;
-}
-
-.bg-l-2-grey {
-    color: #333333 !important;
-    font-weight: 300 !important;
-}
-
-a.badge-link {
-    color: #333333;  
-    text-decoration: none !important;	    
-}
-
-a.badge-link:hover {
-    color: #333333; 
-    text-decoration: none !important;                
-}
-
-a.badge-link:visited {
-    color: #333333;  
-    text-decoration: none !important;              
-}
-
-#dataset-table {
-	font-size: 16px;
-}
-
-tr {
-	height: 60px;
-}
-
-th div {
-    font-weight: 700;
-}
-
-.dataset-list {
-    font-size: 16px;
-    border-bottom: 1px solid #e0e0e0;
-    padding-top: 16px;
-    padding-bottom: 16px;
-    /*margin: -1px -1px 0px -1px;*/    
-}
-
-.dataset-link {
-    color: #333333;  
-    vertical-align: middle;
-}
-
-a.dataset-link:hover {
-    color: #ef5350;
-}  
-
-a.dataset-link:visited {
-    color: #333333; 
-} 
-
-.datasetArrow {
-    cursor: pointer;
-}
-
-.model-header {
-	font-size: 16px !important;
-    color: #333333 !important;	
-}
-
-.modal .modal-content .modal-body {
-	font-size: 16px !important;
-	color: #333333 !important;
-}
-
-.modal-body-paragraph {
-	padding-bottom: 16px;
-}
-
-#modalContainer {
-	overflow-y: auto;
-}
-
-.theme-red .menu:not([class*="bg-"]) .list li.active > :first-child i {
-    color: rgb(116, 116, 116) !important;
-}
