--- src/main/resources/static/index.html
@@ -1,40 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>Geodaten • Kanton Solothurn</title>
-
-    <meta charset="utf-8"/>
-    <meta http-equiv="X-UA-Compatible" content="chrome=1" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-    <meta name="HandheldFriendly" content="true" />
-
-    <script type="text/javascript" language="javascript" src="module1/module1.nocache.js"></script>    
-    <script type="text/javascript" src="//cdn.rawgit.com/openlayers/openlayers.github.io/master/en/v5.3.0/build/ol.js"></script>
-    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/proj4js/2.5.0/proj4.js"></script>    
-
-    <link rel="icon" type="image/png" href="/favicon-32x32.png" sizes="32x32">
-    <link rel="icon" type="image/png" href="/favicon-16x16.png" sizes="16x16">
-    <link rel="icon" type="image/png" href="/android-chrome-192x192.png" sizes="192x192">
-    <link rel="icon" type="image/png" href="/favicon-96x96.png" sizes="96x96">
-    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">    
-    
-    <link type="text/css" rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Fira+Sans:ital,wght@0,400;0,600;0,800;1,400;1,600;1,800&display=swap">
-    <link href="https://fonts.googleapis.com/css2?family=Barlow:wght@300;900&display=swap" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@300;500;700;900&display=swap" rel="stylesheet">
-    
-    <link type="text/css" rel="stylesheet" href="//cdn.rawgit.com/openlayers/openlayers.github.io/master/en/v5.3.0/css/ol.css">    
-    <link type="text/css" rel="stylesheet" href="module1/css/domino-ui.css">
-    <link type="text/css" rel="stylesheet" href="module1/css/themes/all-themes.css">
-    <link type="text/css" rel="stylesheet" href="client.css">
-    <link type="text/css" rel="stylesheet" href="fonts.css">
-  </head>
-
-<body>
-  <noscript>
-    <div style="width: 22em; position: absolute; left: 50%; margin-left: -11em; color: red; background-color: white; border: 1px solid red; padding: 4px; font-family: sans-serif">
-        Your web browser must have JavaScript enabled
-        in order for this application to display correctly.
-    </div>
-  </noscript>
-</body>
-</html>
