--- src/test/java/ch/so/agi/sodata/wgc/BootGwtApplicationTests.java
@@ -1,15 +1,0 @@
-package ch.so.agi.sodata.wgc;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@SpringBootTest
-public class BootGwtApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
